reader/requirements.txt
mysql-connector-python==8.0.33
prometheus-client==0.17.1
flask==2.3.3
werkzeug==2.3.7
